fix/relocate-setup-params
